+++ Request.php
@@ -479,7 +479,7 @@ class Request extends Message
     {
         $parts = [];
         foreach ($this->getCookies() as $cookie) {
-            if (!$this->validateCookieValue($cookie->name)) {
+            if (!$this->validateCookieName($cookie->name)) {
                 throw new InvalidConfigException("Cookie name '{$cookie->name}' is invalid");
             }
             if (!$this->validateCookieValue($cookie->value)) {
@@ -491,15 +491,29 @@ class Request extends Message
     }
 
     /**
-     * Validates cookie name or value.
+     * Validates cookie name.
+     * @param string $name cookie name.
+     * @return bool whether value is valid
+     * @see http://www.ietf.org/rfc/rfc6265.txt
+     * @since 2.0.5
+     */
+    private function validateCookieName($name)
+    {
+        // According to RFC6265 (RFC2616) invalid are: control characters (0-31;127), space, tab and the following: ()<>@,;:\"/?={}'
+        return !preg_match('/[\x00-\x20\x22\x27-\x29\x2c\x2f\x3a-\x40\x5c\x7b\x7d\x7f]/', $name);
+    }
+
+    /**
+     * Validates cookie value.
      * @param string $value cookie value.
      * @return bool whether value is valid
+     * @see http://www.ietf.org/rfc/rfc6265.txt
      * @since 2.0.4
      */
     private function validateCookieValue($value)
     {
-        // Invalid are: control characters (0-31;127), space, tab and the following: ()<>@,;:\"/?={}'
-        return !preg_match('/[\x00-\x20\x22\x28-\x29\x2c\x2f\x3a-\x40\x5c\x7b\x7d\x7f]/', $value);
+        // According to RFC6265 invalid are: control characters (0-31;127), space, tab and the following: ,;\"
+        return !preg_match('/[\x00-\x20\x22\x2c\x3b\x5c\x7f]/', $value);
     }
 
     /**