+++ tests/RequestTest.php
@@ -368,35 +368,75 @@ PART2
     }
 
     /**
-     * Data provider for [[testValidateCookie()]]
+     * Data provider for [[testValidateCookieName()]]
      * @return array test data
      */
-    public function dataProviderValidateCookie()
+    public function dataProviderValidateCookieName()
     {
         return [
             ['foo', true],
             ['utf-абвгд', true],
             [';', false],
+            [',', false],
             ["new\nline", false],
+            ['double"quotes', false],
+            ["single'quotes", false],
             ['/', false],
+            ['x+y=z', false],
             ['\\', false],
+            ['({<brackets>})', false],
+            ['user@domain.com', false],
         ];
     }
 
     /**
-     * @dataProvider dataProviderValidateCookie
+     * @dataProvider dataProviderValidateCookieName
+     *
+     * @param string $name
+     * @param bool $isValid
+     */
+    public function testValidateCookieName($name, $isValid)
+    {
+        $request = new Request();
+        $this->assertEquals($isValid, $this->invoke($request, 'validateCookieName', [$name]));
+    }
+
+    /**
+     * Data provider for [[testValidateCookieValue()]]
+     * @return array test data
+     */
+    public function dataProviderValidateCookieValue()
+    {
+        return [
+            ['foo', true],
+            ['utf-абвгд', true],
+            [';', false],
+            [',', false],
+            ["new\nline", false],
+            ['double"quotes', false],
+            ["single'quotes", true],
+            ['/', true],
+            ['x+y=z', true],
+            ['\\', false],
+            ['({<brackets>})', true],
+            ['user@domain.com', true],
+        ];
+    }
+
+    /**
+     * @dataProvider dataProviderValidateCookieValue
      *
      * @param string $value
      * @param bool $isValid
      */
-    public function testValidateCookie($value, $isValid)
+    public function testValidateCookieValue($value, $isValid)
     {
         $request = new Request();
         $this->assertEquals($isValid, $this->invoke($request, 'validateCookieValue', [$value]));
     }
 
     /**
-     * @depends testValidateCookie
+     * @depends testValidateCookieValue
      */
     public function testComposeCookieHeader()
     {